backend/requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
requests
pillow
googletrans==3.1.0a0